refactor: Consolidate type definitions and functions in types.go
