+++ transport/httptransport/types.go
@@ -1,65 +1,112 @@
 package httptransport
 
 import (
+    "context"
     "encoding/json"
+    "fmt"
+    "net/http"
     "strconv"
 
     "github.com/c0deZ3R0/go-sync-kit/synckit"
     "github.com/c0deZ3R0/go-sync-kit/storage/sqlite"
 )
 
-// JSONEventWithVersion represents an event with version in JSON format
+// JSONEvent is a JSON-serializable representation of an Event
 type JSONEvent struct {
     ID          string                 `json:"id"`
     Type        string                 `json:"type"`
     AggregateID string                 `json:"aggregate_id"`
-    Data        map[string]interface{} `json:"data,omitempty"`
-    Metadata    map[string]interface{} `json:"metadata,omitempty"`
+    Data        interface{}            `json:"data"`
+    Metadata    map[string]interface{} `json:"metadata"`
 }
 
+// JSONEventWithVersion is a JSON-serializable representation of EventWithVersion
 type JSONEventWithVersion struct {
     Event   JSONEvent `json:"event"`
     Version string    `json:"version"`
 }
 
-func toJSONEventWithVersion(ev synckit.EventWithVersion) JSONEventWithVersion {
-    jsonEvent := JSONEvent{
-        ID:          ev.Event.ID(),
-        Type:        ev.Event.Type(),
-        AggregateID: ev.Event.AggregateID(),
-        Data:        ev.Event.Data(),
-        Metadata:    ev.Event.Metadata(),
-    }
+// SimpleEvent is a simple implementation of synckit.Event for HTTP transport
+type SimpleEvent struct {
+    IDValue          string                 `json:"id"`
+    TypeValue        string                 `json:"type"`
+    AggregateIDValue string                 `json:"aggregate_id"`
+    DataValue        interface{}            `json:"data"`
+    MetadataValue    map[string]interface{} `json:"metadata"`
+}
 
-    // Convert version to string
-    version := "0"
-    if iv, ok := ev.Version.(sqlite.IntegerVersion); ok {
-        version = strconv.FormatInt(int64(iv), 10)
-    }
+func (e *SimpleEvent) ID() string                       { return e.IDValue }
+func (e *SimpleEvent) Type() string                     { return e.TypeValue }
+func (e *SimpleEvent) AggregateID() string              { return e.AggregateIDValue }
+func (e *SimpleEvent) Data() interface{}                { return e.DataValue }
+func (e *SimpleEvent) Metadata() map[string]interface{} { return e.MetadataValue }
 
-    return JSONEventWithVersion{
-        Event:   jsonEvent,
-        Version: version,
+// toJSONEvent converts a synckit.Event to JSONEvent
+func toJSONEvent(event synckit.Event) JSONEvent {
+    return JSONEvent{
+        ID:          event.ID(),
+        Type:        event.Type(),
+        AggregateID: event.AggregateID(),
+        Data:        event.Data(),
+        Metadata:    event.Metadata(),
     }
 }
 
-func fromJSONEventWithVersion(je JSONEventWithVersion) synckit.EventWithVersion {
-    ev := &SimpleEvent{
-        IDValue:          je.Event.ID,
-        TypeValue:        je.Event.Type,
-        AggregateIDValue: je.Event.AggregateID,
-        DataValue:        je.Event.Data,
-        MetadataValue:    je.Event.Metadata,
+// toJSONEventWithVersion converts synckit.EventWithVersion to JSONEventWithVersion
+func toJSONEventWithVersion(ev synckit.EventWithVersion) JSONEventWithVersion {
+    return JSONEventWithVersion{
+        Event:   toJSONEvent(ev.Event),
+        Version: ev.Version.String(),
+    }
+}
+
+// fromJSONEvent converts JSONEvent to a concrete Event implementation
+func fromJSONEvent(je JSONEvent) synckit.Event {
+    return &SimpleEvent{
+        IDValue:          je.ID,
+        TypeValue:        je.Type,
+        AggregateIDValue: je.AggregateID,
+        DataValue:        je.Data,
+        MetadataValue:    je.Metadata,
+    }
+}
+
+// fromJSONEventWithVersion converts JSONEventWithVersion back to synckit.EventWithVersion
+// It uses the provided EventStore to parse the version string, making it version-implementation agnostic
+func fromJSONEventWithVersion(ctx context.Context, store synckit.EventStore, jev JSONEventWithVersion) (synckit.EventWithVersion, error) {
+    version, err := store.ParseVersion(ctx, jev.Version)
+    if err != nil {
+        return synckit.EventWithVersion{}, fmt.Errorf("invalid version: %w", err)
     }
 
-    // Parse version as integer
-    var version sqlite.IntegerVersion
-    if v, err := strconv.ParseInt(je.Version, 10, 64); err == nil {
-        version = sqlite.IntegerVersion(v)
+    event := &SimpleEvent{
+        IDValue:          jev.Event.ID,
+        TypeValue:        jev.Event.Type,
+        AggregateIDValue: jev.Event.AggregateID,
+        DataValue:        jev.Event.Data,
+        MetadataValue:    jev.Event.Metadata,
     }
 
     return synckit.EventWithVersion{
-        Event:   ev,
+        Event:   event,
         Version: version,
+    }, nil
+}
+
+// Helper functions for HTTP responses
+func respondWithError(w http.ResponseWriter, code int, message string) {
+    respondWithJSON(w, code, map[string]string{"error": message})
+}
+
+func respondWithJSON(w http.ResponseWriter, code int, payload interface{}) {
+    response, err := json.Marshal(payload)
+    if err != nil {
+        // Fallback if payload marshaling fails
+        w.WriteHeader(http.StatusInternalServerError)
+        w.Write([]byte(`{"error": "failed to marshal response"}`)) 
+        return
     }
+    w.Header().Set("Content-Type", "application/json")
+    w.WriteHeader(code)
+    w.Write(response)
 }
